+++ packages/remix-react/index.tsx
@@ -6,12 +6,16 @@ export type {
   NavigateFunction,
   Params,
   Path,
+  ShouldRevalidateFunction,
   SubmitFunction,
   SubmitOptions,
 } from "react-router-dom";
 export {
   Form,
   Outlet,
+  useAsyncError,
+  useAsyncValue,
+  isRouteErrorResponse,
   useBeforeUnload,
   useFormAction,
   useHref,
@@ -24,23 +28,26 @@ export {
   useParams,
   useResolvedPath,
   useRevalidator,
+  useRouteError,
+  useRouteLoaderData,
   useSearchParams,
   useSubmit,
 } from "react-router-dom";
 
 export type {
+  AwaitProps,
   FetcherWithComponents,
   RouteMatch,
   RemixNavLinkProps as NavLinkProps,
   RemixLinkProps as LinkProps,
 } from "./components";
 export {
+  Await,
   Meta,
   Links,
   Scripts,
   Link,
   NavLink,
-  RemixEntry,
   PrefetchPageLinks,
   LiveReload,
   useTransition,
@@ -49,6 +56,7 @@ export {
   useLoaderData,
   useMatches,
   useActionData,
+  RemixContext as UNSAFE_RemixContext,
 } from "./components";
 
 export type { FormMethod, FormEncType } from "./data";
@@ -59,8 +67,8 @@ export { useCatch } from "./errorBoundaries";
 export type { HtmlLinkDescriptor } from "./links";
 export type {
   HtmlMetaDescriptor,
-  CatchBoundaryComponent,
-  RouteModules,
+  RouteModules as UNSAFE_RouteModules,
+  ShouldReloadFunction,
 } from "./routeModules";
 
 export { ScrollRestoration } from "./scroll-restoration";
@@ -70,6 +78,13 @@ export { RemixServer } from "./server";
 
 export type { Fetcher } from "./transition";
 
-export type { AssetsManifest, EntryContext } from "./entry";
-export type { RouteData } from "./routeData";
-export type { EntryRoute, RouteManifest } from "./routes";
+export type {
+  FutureConfig as UNSAFE_FutureConfig,
+  AssetsManifest as UNSAFE_AssetsManifest,
+  RemixContextObject as UNSAFE_RemixContextObject,
+} from "./entry";
+
+export type {
+  EntryRoute as UNSAFE_EntryRoute,
+  RouteManifest as UNSAFE_RouteManifest,
+} from "./routes";