+++ packages/remix-react/server.tsx
@@ -7,11 +7,16 @@ import {
 
 import { RemixContext } from "./components";
 import type { EntryContext } from "./entry";
+import {
+  RemixErrorBoundary,
+  RemixRootDefaultErrorBoundary,
+} from "./errorBoundaries";
 import { createServerRoutes } from "./routes";
 
 export interface RemixServerProps {
   context: EntryContext;
   url: string | URL;
+  abortDelay?: number;
 }
 
 /**
@@ -19,13 +24,21 @@ export interface RemixServerProps {
  * `app/entry.server.js`). This component is used to generate the HTML in the
  * response from the server.
  */
-export function RemixServer({ context, url }: RemixServerProps): ReactElement {
+export function RemixServer({
+  context,
+  url,
+  abortDelay,
+}: RemixServerProps): ReactElement {
   if (typeof url === "string") {
     url = new URL(url);
   }
 
   let { manifest, routeModules, serverHandoffString } = context;
-  let routes = createServerRoutes(manifest.routes, routeModules);
+  let routes = createServerRoutes(
+    manifest.routes,
+    routeModules,
+    context.future
+  );
   let router = createStaticRouter(routes, context.staticHandlerContext);
 
   return (
@@ -35,13 +48,19 @@ export function RemixServer({ context, url }: RemixServerProps): ReactElement {
         routeModules,
         serverHandoffString,
         future: context.future,
+        abortDelay,
       }}
     >
-      <StaticRouterProvider
-        router={router}
-        context={context.staticHandlerContext}
-        hydrate={false}
-      />
+      <RemixErrorBoundary
+        location={router.state.location}
+        component={RemixRootDefaultErrorBoundary}
+      >
+        <StaticRouterProvider
+          router={router}
+          context={context.staticHandlerContext}
+          hydrate={false}
+        />
+      </RemixErrorBoundary>
     </RemixContext.Provider>
   );
 }