+++ packages/remix-react/CHANGELOG.md
@@ -1,5 +1,23 @@
 # `@remix-run/react`
 
+## 1.10.1
+
+### Patch Changes
+
+- Fetchers should persist data through reload/resubmit ([#5065](https://github.com/remix-run/remix/pull/5065))
+- Update babel config to transpile down to node 14 ([#5047](https://github.com/remix-run/remix/pull/5047))
+
+## 1.10.0
+
+### Minor Changes
+
+- Update Remix to use new data APIs introduced in React Router v6.4 ([#4900](https://github.com/remix-run/remix/pull/4900))
+- Added new hooks from React Router
+  - [`useNavigation`](https://reactrouter.com/en/main/hooks/use-navigation)
+  - [`useNavigationType`](https://reactrouter.com/en/main/hooks/use-navigation-type)
+  - [`useRevalidator`](https://reactrouter.com/en/main/hooks/use-revalidator)
+  - [`useRouteLoaderData`](https://reactrouter.com/en/main/hooks/use-route-loader-data)
+
 ## 1.9.0
 
 ### Patch Changes